+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ## Fixed
 - Fix printing of boolean options in the print-invocations plugin
-- Fix plugin name normalization
+- Breaking! Unqualified plugin names containing dots are no longer
+    normalized to contain the `kaocha.plugin`-namespace in front.
 
 ## Changed