+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Added
 
+- New `gc-profiling` plugin for measuring tests' memory usage.
+
 ## Fixed
 
 - Fix only considering public vars when building up the test plan